дописать

+++ Libr/Data/Models/book.cs
@@ -14,7 +14,7 @@ namespace Libr.Models
         public string Author { get; set; }
         public string description { get; set; }
         public ushort count { get; set; }
-        public ushort price { get; set; }
+        public float price { get; set; }
         public string PhotoPath { get; set; }
         public Genre? genres { get; set; }
     }
